-Wold-style-casting (core & os only (too many warnings in wxWidgets headers))
git-svn-id: svn://svn.code.sf.net/p/passwordsafe/code/branches/x86_64@3654 1f79f812-37fb-46fe-a122-30589dd2bf55

+++ src/corelib/PWScore.h
@@ -260,10 +260,10 @@ public:
 
   // alias/base and shortcut/base handling
   void SortDependents(UUIDVector &dlist, StringX &csDependents);
-  int NumAliases(const uuid_array_t &base_uuid) const
-  {return (int)m_base2aliases_mmap.count(base_uuid);}
-  int NumShortcuts(const uuid_array_t &base_uuid) const
-  {return (int)m_base2shortcuts_mmap.count(base_uuid);}
+  size_t NumAliases(const uuid_array_t &base_uuid) const
+  {return m_base2aliases_mmap.count(base_uuid);}
+  size_t NumShortcuts(const uuid_array_t &base_uuid) const
+  {return m_base2shortcuts_mmap.count(base_uuid);}
 
   ItemListIter GetUniqueBase(const StringX &title, bool &bMultiple);
   ItemListIter GetUniqueBase(const StringX &grouptitle,
